+++ src/Opulence/Framework/Databases/Bootstrappers/MigrationBootstrapper.php
@@ -10,6 +10,8 @@
 
 namespace Opulence\Framework\Databases\Bootstrappers;
 
+use Opulence\Databases\Adapters\Pdo\MySql\Driver as MySqlDriver;
+use Opulence\Databases\Adapters\Pdo\PostgreSql\Driver as PostgreSqlDriver;
 use Opulence\Databases\IConnection;
 use Opulence\Databases\Migrations\FileMigrationFinder;
 use Opulence\Databases\Migrations\IExecutedMigrationRepository;
@@ -22,7 +24,10 @@ use Opulence\Framework\Databases\Migrations\SqlExecutedMigrationRepository;
 use Opulence\Ioc\Bootstrappers\Bootstrapper;
 use Opulence\Ioc\Bootstrappers\ILazyBootstrapper;
 use Opulence\Ioc\IContainer;
-use Opulence\QueryBuilders\PostgreSql\QueryBuilder;
+use Opulence\QueryBuilders\MySql\QueryBuilder as MySqlQueryBuilder;
+use Opulence\QueryBuilders\PostgreSql\QueryBuilder as PostgreSqlQueryBuilder;
+use Opulence\QueryBuilders\QueryBuilder as BaseQueryBuilder;
+use RuntimeException;
 
 /**
  * Defines the database migration bootstrapper
@@ -57,13 +62,33 @@ class MigrationBootstrapper extends Bootstrapper implements ILazyBootstrapper
      *
      * @param IContainer $container The IoC container
      * @return IExecutedMigrationRepository The executed migration repository
+     * @throws RuntimeException Thrown if there was an error resolving the query builder
      */
     protected function getExecutedMigrationRepository(IContainer $container) : IExecutedMigrationRepository
     {
+        $driverClass = getenv('DB_DRIVER') ?: PostgreSqlDriver::class;
+
+        switch ($driverClass) {
+            case MySqlDriver::class:
+                $queryBuilder = new MySqlQueryBuilder();
+                $container->bindInstance(MySqlQueryBuilder::class, $queryBuilder);
+                break;
+            case PostgreSqlDriver::class:
+                $queryBuilder = new PostgreSqlQueryBuilder();
+                $container->bindInstance(PostgreSqlQueryBuilder::class, $queryBuilder);
+                break;
+            default:
+                throw new RuntimeException(
+                    "Invalid database driver type specified in environment var \"DB_DRIVER\": $driverClass"
+                );
+        }
+
+        $container->bindInstance(BaseQueryBuilder::class, $queryBuilder);
+
         return new SqlExecutedMigrationRepository(
             SqlExecutedMigrationRepository::DEFAULT_TABLE_NAME,
             $container->resolve(IConnection::class),
-            new QueryBuilder(),
+            $queryBuilder,
             new TypeMapperFactory()
         );
     }